impedancegpt_final/rag_vlm/requirements.txt.txt
faiss-cpu
sentence-transformers
numpy
pillow
transformers
